Pretty complete code to identify every allocation.
Code that, for every struct type generates a verification function that 
verifies each member.  All members are verified by the go_verify macro, 
which is an expression that only compiles for those members that we care 
about.  So each function completely describes the associated struct, but 
doesn't generate any code for non-checked members.


git-svn-id: svn+ssh://plant/filer/osa/svn-hype/osck@10 0a08b987-4bf4-4c34-b5fc-5002ad2b9887